Centralize GraphQL response helpers under service/tests/common/graphql.rs
Extract shared GraphQL testing utilities from graphql_tests.rs into a
reusable module:
- execute_query() - Run GraphQL queries against test schema
- extract_data() - Get response data field
- extract_errors() - Get response errors array
- assert_no_errors() / assert_has_errors() - Response assertions

This follows the existing pattern of common test utilities like
app_builder and test_db.
